rename sidebar and fix scroll
