+++ app/frontend/components/domains/permit-project/index.tsx
@@ -8,8 +8,8 @@ import { useSearch } from "../../../hooks/use-search"
 import { useMst } from "../../../setup/root"
 import { LoadingScreen } from "../../shared/base/loading-screen"
 import { ComingSoonPlaceholder } from "../navigation/coming-soon-placeholder"
+import { ITabItem, ProjectSidebarTabList } from "./project-sidebar-tab-list"
 import { ProjectTabPanelContent } from "./project-tab-panel-content"
-import { ITabItem, ProjectSidebarTabList } from "./sidebar-tab-list"
 import { StepCodeTabPanelContent } from "./step-code-tab-panel-content"
 
 interface IProjectDashboardScreenProps {}
